Delegate topology hint gen to CPU manager policy
- The previous implementation depended on a fixed set of policies.
