+++ pkg/kubelet/cm/cpumanager/policy_none.go
@@ -20,6 +20,7 @@ import (
 	"k8s.io/api/core/v1"
 	"k8s.io/klog"
 	"k8s.io/kubernetes/pkg/kubelet/cm/cpumanager/state"
+	"k8s.io/kubernetes/pkg/kubelet/cm/topologymanager"
 )
 
 type nonePolicy struct{}
@@ -47,5 +48,9 @@ func (p *nonePolicy) AddContainer(s state.State, pod *v1.Pod, container *v1.Cont
 }
 
 func (p *nonePolicy) RemoveContainer(s state.State, containerID string) error {
+	return nil
+}
+
+func (p *nonePolicy) GetTopologyHints(s state.State, pod v1.Pod, container v1.Container) map[string][]topologymanager.TopologyHint {
 	return nil
 }
